benchmarks/02_training/requirements.txt
# Requirements for Classical ML Training Pipeline

# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Chemistry and Molecular Featurization
rdkit>=2022.03.1

# Configuration and Utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Model Persistence
joblib>=1.1.0

# Experiment Tracking (Optional)
# wandb>=0.15.0  # For experiment tracking and visualization

# Optional: Faster GBM backends (GBMTrainer falls back to sklearn HistGBM if absent)
# xgboost>=1.5.0
# lightgbm>=3.3.0
